# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FlagLab API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plays:
    get:
      operationId: listPlays
      tags: [plays]
      summary: List all plays
      parameters:
        - name: playbook_id
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [5v5, 7v7]
        - name: mode
          in: query
          schema:
            type: string
            enum: [offense, defense]
        - name: field_zone
          in: query
          schema:
            type: string
            enum: [red_zone, midfield, backed_up]
        - name: search
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of plays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Play"
    post:
      operationId: createPlay
      tags: [plays]
      summary: Create a play
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayBody"
      responses:
        "201":
          description: Created play
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Play"
  /plays/{id}:
    get:
      operationId: getPlay
      tags: [plays]
      summary: Get a play by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Play found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Play"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePlay
      tags: [plays]
      summary: Update a play
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayBody"
      responses:
        "200":
          description: Updated play
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Play"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePlay
      tags: [plays]
      summary: Delete a play
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /plays/{id}/duplicate:
    post:
      operationId: duplicatePlay
      tags: [plays]
      summary: Duplicate a play
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Duplicated play
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Play"
  /playbooks:
    get:
      operationId: listPlaybooks
      tags: [playbooks]
      summary: List all playbooks
      responses:
        "200":
          description: List of playbooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playbook"
    post:
      operationId: createPlaybook
      tags: [playbooks]
      summary: Create a playbook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaybookBody"
      responses:
        "201":
          description: Created playbook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playbook"
  /playbooks/{id}:
    get:
      operationId: getPlaybook
      tags: [playbooks]
      summary: Get a playbook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Playbook found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playbook"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePlaybook
      tags: [playbooks]
      summary: Update a playbook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlaybookBody"
      responses:
        "200":
          description: Updated playbook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playbook"
    delete:
      operationId: deletePlaybook
      tags: [playbooks]
      summary: Delete a playbook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    PlayerToken:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          description: "QB, WR1, WR2, WR3, RB, C, DB, Rusher, Safety, LB, etc."
        label:
          type: string
        x:
          type: number
        y:
          type: number
        color:
          type: string
        team:
          type: string
          enum: [offense, defense]
      required:
        - id
        - role
        - label
        - x
        - y
        - team
    RoutePoint:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
      required:
        - x
        - y
    RouteSegment:
      type: object
      properties:
        id:
          type: string
        playerId:
          type: string
        type:
          type: string
          enum: [straight, slant, out, in, post, corner, go, comeback, hitch, drag, wheel, motion, handoff, zone_drop, screen]
        points:
          type: array
          items:
            $ref: "#/components/schemas/RoutePoint"
        style:
          type: string
          enum: [solid, dashed, dotted]
        color:
          type: string
      required:
        - id
        - playerId
        - type
        - points
        - style
    Play:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        mode:
          type: string
          enum: [offense, defense]
        format:
          type: string
          enum: [5v5, 7v7]
        formationId:
          type: string
        playbookId:
          type: string
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerToken"
        routes:
          type: array
          items:
            $ref: "#/components/schemas/RouteSegment"
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        suggestedUse:
          type: string
        coverageTargets:
          type: array
          items:
            type: string
        fieldZone:
          type: string
          enum: [red_zone, midfield, backed_up]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        primaryConcept:
          type: string
        isManBeater:
          type: boolean
        isZoneBeater:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - mode
        - format
        - players
        - routes
        - tags
        - coverageTargets
        - createdAt
        - updatedAt
    CreatePlayBody:
      type: object
      properties:
        title:
          type: string
        mode:
          type: string
          enum: [offense, defense]
        format:
          type: string
          enum: [5v5, 7v7]
        formationId:
          type: string
        playbookId:
          type: string
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerToken"
        routes:
          type: array
          items:
            $ref: "#/components/schemas/RouteSegment"
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        suggestedUse:
          type: string
        coverageTargets:
          type: array
          items:
            type: string
        fieldZone:
          type: string
          enum: [red_zone, midfield, backed_up]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        primaryConcept:
          type: string
        isManBeater:
          type: boolean
        isZoneBeater:
          type: boolean
      required:
        - title
        - mode
        - format
    UpdatePlayBody:
      type: object
      properties:
        title:
          type: string
        mode:
          type: string
          enum: [offense, defense]
        format:
          type: string
          enum: [5v5, 7v7]
        formationId:
          type: string
        playbookId:
          type: string
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerToken"
        routes:
          type: array
          items:
            $ref: "#/components/schemas/RouteSegment"
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        suggestedUse:
          type: string
        coverageTargets:
          type: array
          items:
            type: string
        fieldZone:
          type: string
          enum: [red_zone, midfield, backed_up]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        primaryConcept:
          type: string
        isManBeater:
          type: boolean
        isZoneBeater:
          type: boolean
    Playbook:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        format:
          type: string
          enum: [5v5, 7v7, both]
        color:
          type: string
        playCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - format
        - playCount
        - createdAt
        - updatedAt
    CreatePlaybookBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        format:
          type: string
          enum: [5v5, 7v7, both]
        color:
          type: string
      required:
        - name
        - format
    UpdatePlaybookBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        format:
          type: string
          enum: [5v5, 7v7, both]
        color:
          type: string
...