# lib/api-zod/src/generated/types/routeSegment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlagLab API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoutePoint } from "./routePoint";
import type { RouteSegmentStyle } from "./routeSegmentStyle";
import type { RouteSegmentType } from "./routeSegmentType";

export interface RouteSegment {
  id: string;
  playerId: string;
  type: RouteSegmentType;
  points: RoutePoint[];
  style: RouteSegmentStyle;
  color?: string;
}