# lib/api-zod/src/generated/types/updatePlaybookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlagLab API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePlaybookBodyFormat } from "./updatePlaybookBodyFormat";

export interface UpdatePlaybookBody {
  name?: string;
  description?: string;
  format?: UpdatePlaybookBodyFormat;
  color?: string;
}